requirements.txt
fastapi==0.115.0
psycopg2-binary==2.9.3
python-dotenv==1.0.1
SQLAlchemy==2.0.35
faker==30.3.0
Jinja2==3.1.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0